Fix : fixed a npe with the new way maxhp is calculated.

+++ L2_GameServer_c5/java/net/sf/l2j/gameserver/model/L2Attackable.java
@@ -1278,7 +1278,7 @@ public class L2Attackable extends L2NpcInstance
     	 }
          
          //Instant Item Drop :>
-    	 int rateHp = (int) getStat().calcStat(Stats.MAX_HP , 1, this, null);    	 
+    	 double rateHp = getStat().calcStat(Stats.MAX_HP , 1, this, null);
          if (rateHp <= 1 && String.valueOf(npcTemplate.type).contentEquals("L2Monster")) //only L2Monster with <= 1x HP can drop herbs
          {
         	 boolean _hp = false;
