+++ L2_GameServer_c5/java/net/sf/l2j/gameserver/model/actor/instance/L2NpcInstance.java
@@ -637,7 +637,7 @@ public class L2NpcInstance extends L2Character
             html1.append("<font color=\"LEVEL\">Combat</font>");
             html1.append("<table border=\"0\" width=\"100%\">");
             html1.append("<tr><td>Current HP</td><td>"+getCurrentHp()+"</td><td>Current MP</td><td>"+getCurrentMp()+"</td></tr>");
-            html1.append("<tr><td>Max.HP</td><td>"+(int)(getMaxHp()/(int) getStat().calcStat(Stats.MAX_HP , 1, this, null))+"*"+(int) getStat().calcStat(Stats.MAX_HP , 1, this, null)+"</td><td>Max.MP</td><td>"+getMaxMp()+"</td></tr>");
+            html1.append("<tr><td>Max.HP</td><td>"+(int)(getMaxHp()/getStat().calcStat(Stats.MAX_HP , 1, this, null))+"*"+getStat().calcStat(Stats.MAX_HP , 1, this, null)+"</td><td>Max.MP</td><td>"+getMaxMp()+"</td></tr>");
             html1.append("<tr><td>P.Atk.</td><td>"+getPAtk(null)+"</td><td>M.Atk.</td><td>"+getMAtk(null,null)+"</td></tr>");
             html1.append("<tr><td>P.Def.</td><td>"+getPDef(null)+"</td><td>M.Def.</td><td>"+getMDef(null,null)+"</td></tr>");
             html1.append("<tr><td>Accuracy</td><td>"+getAccuracy()+"</td><td>Evasion</td><td>"+getEvasionRate(null)+"</td></tr>");
@@ -2089,7 +2089,7 @@ public class L2NpcInstance extends L2Character
      */
     public int getExpReward()
     {
-    	int rateXp = (int) getStat().calcStat(Stats.MAX_HP , 1, this, null);
+    	double rateXp = getStat().calcStat(Stats.MAX_HP , 1, this, null);
         return (int)(getTemplate().rewardExp * rateXp * Config.RATE_XP);
     }
     
@@ -2098,7 +2098,7 @@ public class L2NpcInstance extends L2Character
      */
     public int getSpReward()
     {
-    	int rateSp = (int) getStat().calcStat(Stats.MAX_HP , 1, this, null);
+    	double rateSp = getStat().calcStat(Stats.MAX_HP , 1, this, null);
         return (int)(getTemplate().rewardSp * rateSp * Config.RATE_SP);
     }
